+++ CHANGELOG.md
@@ -7,7 +7,10 @@ Package-specific changes not released in any SDK will be added here just before 
 
 ### 📚 3rd party library updates
 
+
+- Updated `lottie-react-native` from `3.5.0` to `4.0.2`. ([#13151](https://github.com/expo/expo/pull/13151) by [@tsapeta](https://github.com/tsapeta))
 - Updated `@react-native-community/picker@1.9.11` to `@react-native-picker/picker@1.16.1`. ([#13152](https://github.com/expo/expo/pull/13152) by [@ajsmth](https://github.com/ajsmth))
+
 
 ### 🛠 Breaking changes